Add WorkflowsTests to system tests
This commit adds the WorkflowsTests class to the system tests in order to include tests for workflows functionality. The WorkflowsTests class is imported in the tests.py file and the relevant test methods are added to the list of all tests. This change allows for testing the workflows feature in the system tests.